Replaced Evil Insta-Checkbox with small properties dialog for designating JMS endpoints as monitored message sources.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@5359 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ src/com/l7tech/console/panels/JmsEndpointsWindow.java
@@ -12,6 +12,8 @@ import com.l7tech.common.transport.jms.JmsEndpoint;
 import com.l7tech.console.util.Registry;
 
 import javax.swing.*;
+import javax.swing.event.ListSelectionListener;
+import javax.swing.event.ListSelectionEvent;
 import javax.swing.table.AbstractTableModel;
 import java.awt.*;
 import java.awt.event.ActionEvent;
@@ -27,8 +29,8 @@ import java.util.ArrayList;
 public class JmsEndpointsWindow extends JDialog {
     public static final int MESSAGE_SOURCE_COL = 2;
 
-    private JPanel bottomButtonPanel;
     private JButton closeButton;
+    private JButton propertiesButton;
     private JTable endpointTable;
     private EndpointTableModel endpointTableModel;
     private JPanel sideButtonPanel;
@@ -49,6 +51,7 @@ public class JmsEndpointsWindow extends JDialog {
         JScrollPane sp = new JScrollPane(getEndpointTable(),
                                          JScrollPane.VERTICAL_SCROLLBAR_ALWAYS,
                                          JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED);
+        sp.setPreferredSize(new Dimension(400, 200));
         p.add(sp,
               new GridBagConstraints(0, 1, 1, 1, 10.0, 10.0,
                                      GridBagConstraints.CENTER,
@@ -56,18 +59,13 @@ public class JmsEndpointsWindow extends JDialog {
                                      new Insets(5, 5, 5, 5), 0, 0));
 
         p.add(getSideButtonPanel(),
-              new GridBagConstraints(1, 1, 1, 1, 10.0, 10.0,
-                                     GridBagConstraints.CENTER,
-                                     GridBagConstraints.BOTH,
-                                     new Insets(5, 5, 5, 5), 0, 0));
-
-        p.add(getBottomButtonPanel(),
-              new GridBagConstraints(0, 2, 2, 1, 0.0, 0.0,
-                                     GridBagConstraints.EAST,
-                                     GridBagConstraints.NONE,
+              new GridBagConstraints(1, 1, 1, GridBagConstraints.REMAINDER, 0.0, 1.0,
+                                     GridBagConstraints.NORTH,
+                                     GridBagConstraints.VERTICAL,
                                      new Insets(5, 5, 5, 5), 0, 0));
 
         pack();
+        enableOrDisableButtons();
 
     }
 
@@ -91,25 +89,6 @@ public class JmsEndpointsWindow extends JDialog {
             return getEndpointListItems().size();
         }
 
-        public Class getColumnClass(int columnIndex) {
-            return columnIndex == MESSAGE_SOURCE_COL ? Boolean.class : String.class;
-        }
-
-        public void setValueAt(Object aValue, int rowIndex, int columnIndex) {
-            if (columnIndex == MESSAGE_SOURCE_COL && aValue instanceof Boolean) {
-                EndpointListItem i = (EndpointListItem) getEndpointListItems().get(rowIndex);
-                if (i.endpoint != null) {
-                    try {
-                        final boolean b = ((Boolean)aValue).booleanValue();
-                        Registry.getDefault().getJmsManager().setEndpointMessageSource(i.endpoint.getOid(), b);
-                        i.endpoint.setMessageSource(b);
-                    } catch (Exception e) {
-                        throw new RuntimeException("Unable to save changes to endpoint " + i.endpoint, e);
-                    }
-                }
-            }
-        }
-
         public String getColumnName(int column) {
             switch (column) {
                 case 0:
@@ -117,13 +96,9 @@ public class JmsEndpointsWindow extends JDialog {
                 case 1:
                     return "Endpoint";
                 case MESSAGE_SOURCE_COL:
-                    return "Message Source";
+                    return "Direction";
             }
             return "?";
-        }
-
-        public boolean isCellEditable(int rowIndex, int columnIndex) {
-            return columnIndex == MESSAGE_SOURCE_COL;
         }
 
         public Object getValueAt(int rowIndex, int columnIndex) {
@@ -134,7 +109,11 @@ public class JmsEndpointsWindow extends JDialog {
                 case 1:
                     return i.endpoint == null ? "(no endpoints)" : i.endpoint.getName();
                 case MESSAGE_SOURCE_COL:
-                    return i.endpoint == null ? null : new Boolean(i.endpoint.isMessageSource());
+                    if (i.endpoint == null) {
+                        return "";
+                    } else {
+                        return i.endpoint.isMessageSource() ? "Inbound (Monitored)" : "Outbound from Gateway";
+                    }
             }
             return "?";
         }
@@ -143,7 +122,7 @@ public class JmsEndpointsWindow extends JDialog {
             return endpointListItems;
         }
 
-        /** Get the current list of monitored JmsEndpoint EntityHeaders from the server. */
+        /** Get the current list of JmsEndpoints from the server. */
         private ArrayList loadEndpointListItems() {
             try {
                 ArrayList endpointListItems = new ArrayList();
@@ -162,7 +141,7 @@ public class JmsEndpointsWindow extends JDialog {
                 }
                 return endpointListItems;
             } catch (Exception e) {
-                throw new RuntimeException("Unable to look up list of monitored JMS endpoints", e);
+                throw new RuntimeException("Unable to look up list of known JMS endpoints", e);
             }
         }
 
@@ -182,13 +161,26 @@ public class JmsEndpointsWindow extends JDialog {
                                                        GridBagConstraints.CENTER,
                                                        GridBagConstraints.HORIZONTAL,
                                                        new Insets(0, 0, 0, 0), 0, 0));
+            sideButtonPanel.add(getPropertiesButton(),
+                                new GridBagConstraints(0, 2, 1, 1, 0.0, 0.0,
+                                                       GridBagConstraints.CENTER,
+                                                       GridBagConstraints.HORIZONTAL,
+                                                       new Insets(6, 0, 0, 0), 0, 0));
             sideButtonPanel.add(Box.createGlue(),
-                                new GridBagConstraints(0, 2, 1, 1, 1.0, 1.0,
+                                new GridBagConstraints(0, 3, 1, 1, 1.0, 1.0,
                                                        GridBagConstraints.CENTER,
                                                        GridBagConstraints.VERTICAL,
                                                        new Insets(0, 0, 0, 0), 0, 0));
+            sideButtonPanel.add(getCloseButton(),
+                                new GridBagConstraints(0, 4, 1, 1, 0.0, 0.0,
+                                                       GridBagConstraints.SOUTH,
+                                                       GridBagConstraints.HORIZONTAL,
+                                                       new Insets(0, 0, 0, 0), 0, 0));
 
-            Utilities.equalizeButtonSizes(new JButton[] { getAddButton(), getRemoveButton() });
+            Utilities.equalizeButtonSizes(new JButton[] { getAddButton(),
+                                                          getRemoveButton(),
+                                                          getPropertiesButton(),
+                                                          getCloseButton() });
         }
         return sideButtonPanel;
     }
@@ -204,13 +196,25 @@ public class JmsEndpointsWindow extends JDialog {
                         if (i != null) {
                             if (i.endpoint == null)
                                 try {
+                                    // Won't bother with confirmation screen for removing connection with no endpoints
                                     Registry.getDefault().getJmsManager().deleteConnection(i.connection.getOid());
                                 } catch (Exception e1) {
                                     throw new RuntimeException("Unable to delete connection " + i.connection, e1);
                                 }
                             else
                                 try {
-                                    Registry.getDefault().getJmsManager().deleteEndpoint(i.endpoint.getOid());
+                                    Object[] options = { "Remove", "Cancel" };
+                                    int result = JOptionPane.showOptionDialog(null,
+                                                                              "<HTML>Are you sure you want to remove the " +
+                                                                              "registration for the JMS Endpoint " +
+                                                                              i.endpoint + "?<br>" +
+                                                                              "<center>This action cannot be undone." +
+                                                                              "</center></html>",
+                                                                              "Remove Endpoint?",
+                                                                              0, JOptionPane.WARNING_MESSAGE,
+                                                                              null, options, options[1]);
+                                    if (result == 0)
+                                        Registry.getDefault().getJmsManager().deleteEndpoint(i.endpoint.getOid());
                                 } catch (Exception e1) {
                                     throw new RuntimeException("Unable to delete endpoint " + i.endpoint, e1);
                                 }
@@ -236,24 +240,6 @@ public class JmsEndpointsWindow extends JDialog {
             });
         }
         return addButton;
-    }
-
-    private JPanel getBottomButtonPanel() {
-        if (bottomButtonPanel == null) {
-            bottomButtonPanel = new JPanel();
-            bottomButtonPanel.setLayout(new GridBagLayout());
-            bottomButtonPanel.add(Box.createGlue(),
-                                  new GridBagConstraints(0, 0, 1, 1, 1.0, 1.0,
-                                                         GridBagConstraints.CENTER,
-                                                         GridBagConstraints.HORIZONTAL,
-                                                         new Insets(0, 0, 0, 0), 0, 0));
-            bottomButtonPanel.add(getCloseButton(),
-                                  new GridBagConstraints(1, 0, 1, 1, 1.0, 1.0,
-                                                         GridBagConstraints.EAST,
-                                                         GridBagConstraints.NONE,
-                                                         new Insets(0, 0, 0, 5), 0, 0));
-        }
-        return bottomButtonPanel;
     }
 
     private class AddEndpointWindow extends JDialog {
@@ -351,12 +337,52 @@ public class JmsEndpointsWindow extends JDialog {
         return closeButton;
     }
 
+    private JButton getPropertiesButton() {
+        if (propertiesButton == null) {
+            propertiesButton = new JButton("Properties...");
+            propertiesButton.addActionListener(new ActionListener() {
+                public void actionPerformed(ActionEvent e) {
+                    int row = getEndpointTable().getSelectedRow();
+                    if (row >= 0) {
+                        EndpointListItem i = (EndpointListItem) getEndpointTableModel().getEndpointListItems().get(row);
+                        if (i != null && i.endpoint != null) {
+                            new SimpleJmsEndpointPropertiesDialog(i.endpoint).show();
+                            getEndpointTableModel().fireTableRowsUpdated(row, row);
+                        }
+                    }
+                }
+            });
+        }
+        return propertiesButton;
+    }
+
     private JTable getEndpointTable() {
         if (endpointTable == null) {
             endpointTable = new JTable(getEndpointTableModel());
             endpointTable.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
+            endpointTable.getSelectionModel().addListSelectionListener(new ListSelectionListener() {
+                public void valueChanged(ListSelectionEvent e) {
+                    enableOrDisableButtons();
+                }
+            });
         }
         return endpointTable;
+    }
+
+    private void enableOrDisableButtons() {
+        boolean propsEnabled = false;
+        boolean removeEnabled = false;
+        int row = getEndpointTable().getSelectedRow();
+        if (row >= 0) {
+            EndpointListItem i = (EndpointListItem) getEndpointTableModel().getEndpointListItems().get(row);
+            if (i != null) {
+                removeEnabled = true;
+                if (i.endpoint != null)
+                    propsEnabled = true;
+            }
+        }
+        getRemoveButton().setEnabled(removeEnabled);
+        getPropertiesButton().setEnabled(propsEnabled);
     }
 
     private EndpointTableModel getEndpointTableModel() {
@@ -387,6 +413,98 @@ public class JmsEndpointsWindow extends JDialog {
                     getEndpointTable().getSelectionModel().setSelectionInterval(i, i);
                     break;
                 }
+            }
+        }
+    }
+
+    private class SimpleJmsEndpointPropertiesDialog extends JDialog {
+        private JmsEndpoint endpoint;
+
+        SimpleJmsEndpointPropertiesDialog(JmsEndpoint endpoint) {
+            super(JmsEndpointsWindow.this, "Endpoint Properties", true);
+            this.endpoint = endpoint;
+
+            JButton okButton = new JButton("Ok");
+            JButton cancelButton = new JButton("Cancel");
+            JRadioButton outboundButton = new JRadioButton("Outbound");
+            final JRadioButton inboundButton = new JRadioButton("Inbound (Gateway will take messages from this Endpoint)");
+            ButtonGroup buttonGroup = new ButtonGroup();
+            buttonGroup.add(inboundButton);
+            buttonGroup.add(outboundButton);
+            inboundButton.setSelected(endpoint.isMessageSource());
+            outboundButton.setSelected(!inboundButton.isSelected());
+
+            cancelButton.addActionListener(new ActionListener() {
+                public void actionPerformed(ActionEvent e) {
+                    SimpleJmsEndpointPropertiesDialog.this.hide();
+                    SimpleJmsEndpointPropertiesDialog.this.dispose();
+                }
+            });
+            okButton.addActionListener(new ActionListener() {
+                public void actionPerformed(ActionEvent e) {
+                    saveMessageSourceValue(inboundButton.isSelected());
+                    SimpleJmsEndpointPropertiesDialog.this.hide();
+                    SimpleJmsEndpointPropertiesDialog.this.dispose();
+                }
+            });
+
+            Container p = getContentPane();
+            p.setLayout(new GridBagLayout());
+
+            p.add(new JLabel("Name:"),
+                  new GridBagConstraints(0, 0, 1, 1, 0.0, 0.0,
+                                         GridBagConstraints.WEST,
+                                         GridBagConstraints.NONE,
+                                         new Insets(12, 12, 0, 0), 0, 0));
+            p.add(new JLabel(endpoint.getName()),
+                  new GridBagConstraints(1, 0, GridBagConstraints.REMAINDER, 1, 1.0, 0.0,
+                                         GridBagConstraints.WEST,
+                                         GridBagConstraints.HORIZONTAL,
+                                         new Insets(12, 5, 0, 5), 0, 0));
+
+            p.add(new JLabel("Direction:"),
+                  new GridBagConstraints(0, 1, GridBagConstraints.REMAINDER, 1, 0.0, 0.0,
+                                         GridBagConstraints.WEST,
+                                         GridBagConstraints.NONE,
+                                         new Insets(12, 12, 0, 0), 0, 0));
+            p.add(outboundButton,
+                  new GridBagConstraints(0, 2, GridBagConstraints.REMAINDER, 1, 1.0, 1.0,
+                                         GridBagConstraints.WEST,
+                                         GridBagConstraints.NONE,
+                                         new Insets(0, 30, 0, 11), 0, 0));
+            p.add(inboundButton,
+                  new GridBagConstraints(0, 3, GridBagConstraints.REMAINDER, 1, 1.0, 1.0,
+                                         GridBagConstraints.WEST,
+                                         GridBagConstraints.NONE,
+                                         new Insets(0, 30, 11, 11), 0, 0));
+            p.add(okButton,
+                  new GridBagConstraints(3, 4, 1, 1, 0.0, 0.0,
+                                         GridBagConstraints.EAST,
+                                         GridBagConstraints.NONE,
+                                         new Insets(0, 0, 5, 5), 0, 0));
+            p.add(cancelButton,
+                  new GridBagConstraints(4, 4, 1, 1, 0.0, 0.0,
+                                         GridBagConstraints.EAST,
+                                         GridBagConstraints.NONE,
+                                         new Insets(0, 0, 5, 11), 0, 0));
+            p.add(Box.createGlue(),
+                  new GridBagConstraints(1, 4, 1, 1, 1.0, 1.0,
+                                         GridBagConstraints.EAST,
+                                         GridBagConstraints.HORIZONTAL,
+                                         new Insets(0, 0, 0, 0), 0, 0));
+
+            Utilities.equalizeButtonSizes(new JButton[] { okButton, cancelButton });
+
+            pack();
+            Utilities.centerOnScreen(this);
+        }
+
+        public void saveMessageSourceValue(boolean isMessageSource) {
+            try {
+                Registry.getDefault().getJmsManager().setEndpointMessageSource(endpoint.getOid(), isMessageSource);
+                endpoint.setMessageSource(isMessageSource);
+            } catch (Exception e) {
+                throw new RuntimeException("Unable to save changes to endpoint " + endpoint, e);
             }
         }
     }
